fix(infer): pre-scan raw JSON for float-typed fields
JS's JSON.parse collapses "1.0" and "1" into the same Number, so
inference can't recover whether a field's source JSON used decimal
syntax. The masterfile actually uses decimals for ~264 fields
(probabilities, ratios, percentages, intervals) — fields that the
old inference saw integral observations of and labeled `u64`/`i64`,
producing types that serde then refused to deserialize against
real `1.0`-shaped data.

Adds a single regex pass over the raw masterfile JSON text in
src/generate.ts before inference runs. Captures field names
followed by numbers with `.X` or `eX` syntax, populates a
ReadonlySet, then biases inferObjectType to mark matching fields as
`float` regardless of the per-value-derived numericKind.

Side effect (good): parse benchmark went 150ms → 55ms because the
auto-derive untagged dispatch no longer falls back to slow
stub-matching for entries whose payload had type-mismatched
numerics. Now it dispatches correctly on the first try. ~5×
total speedup over the original alphabetical/inference-incorrect
baseline.

False positives in the field-name set (legitimately-int fields that
share a name with a float field elsewhere) are harmless: f64
represents integers up to 2^53 exactly, well above any masterfile
value. False negatives (a float field that never appears with
decimal syntax in the corpus) would still produce u64/i64 — but if
the corpus never showed a decimal, the type matches the data.

401 fields now `f64`, 522 still `u64`, 16 still `i64`. Sensible
distribution.

+++ packages/rust/src/client_quest_template.rs
@@ -18,7 +18,7 @@ pub struct WithCombatType {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct WithDistance {
-    pub distance_km: u64,
+    pub distance_km: f64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
